# product-service/openapi.yml
openapi: 3.0.0
servers:
  - url: https://2to564lyf9.execute-api.eu-west-1.amazonaws.com/dev
components:
  schemas:
    BaseProduct:
      type: object
      required:
        - title
        - price
        - count
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        count:
          type: number
        images:
          type: array
          items:
            type: string
            format: url
    CreateProduct:
      allOf:
        - $ref: '#/components/schemas/BaseProduct'
    Product:
      allOf:
        - type: object
          required:
            - id
          properties:
            id:
              type: string
        - $ref: '#/components/schemas/BaseProduct'
    GetProductListResponse:
      type: array
      items:
        $ref: '#/components/schemas/Product'
info:
  title: RS School App Products Service
  version: '1'
paths:
  /products:
    get:
      tags:
        - Product Service
      operationId: getProductList
      summary: Get Product List
      description: The endpoint returns a list of products.
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductListResponse'
    post:
      tags:
        - Product Service
      operationId: createProduct
      summary: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProduct'
      responses:
        '200':
          description: Newly created product with id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  '/products/{productId}':
    get:
      tags:
        - Product Service
      operationId: getProductById
      parameters:
        - in: path
          name: productId
          schema:
            type: string
          required: true
          description: Product Id
      summary: Get Product By Id
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product Not Found